♻️ Refactor : picsel book & my picsel
