+++ src/feature/picsel/shared/components/ui/molecules/PhotoCard.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { Image, Pressable, Text, View } from 'react-native';
 
-import { formatDate } from '../../../utils/dateUtils';
+import { formatDate } from '../../../../myPicsel/utils/dateUtils';
 
 import SparkleImages from '@/shared/images/Sparkle';
