+++ src/feature/picsel/shared/hooks/animation/useSelectingMode.ts
