+++ src/feature/picsel/shared/hooks/photo/useImageDimensions.ts
