+++ src/feature/picsel/shared/hooks/photo/usePhotoSelection.ts
