+++ src/screens/picsel/picselBook/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import PicselBookTabContent from '@/feature/picsel/shared/components/layouts/PicselBookTabContent';
 import PicselBookTabHeader from '@/feature/picsel/shared/components/layouts/PicselBookTabHeader';
-import { usePicselBookTab } from '@/feature/picsel/shared/hooks/usePicselBookTab';
+import { usePicselBookTab } from '@/feature/picsel/shared/hooks/animation/usePicselBookTab';
 import ScreenLayout from '@/shared/components/layouts/ScreenLayout';
 
 const PicselBookScreen = () => {
